Clear etag on ping exception
Clear etag when an exception is thrown during the ping phase
to prevent an action is ignored.
For example if a target loses the network while it performs an action,
it will try to make a request with the same ETag value, this will cause
the server to respond with a 304 status code, even though the action
was not performed.

Reported by "Sergey Sedoy" (https://github.com/Kynetics/uf-ddiclient/issues/1)

Signed-off-by: Daniele Sergio <daniele.sergio@kynetics.com>